Move CPU memory into Cpu class
The changes merge Memory functionality into Cpu and relocate Main.cpp.
